Decouple debugger and k8s.Controller
Moves getPodForProxy to compute client as unexported method.
Instead, debugger package will now call GetProxyConfig from
compute client, which performs a portforward and returns
the envoy config. This, as well as other methods
moved from k8s and debugger to compute do not have functional changes,
aside from getEnvoyConfig - which now uses the k8s portforward
method instead of its own portforward method.
