+++ pkg/catalog/mock_catalog_generated.go
@@ -20,6 +20,7 @@ import (
 	v1alpha4 "github.com/servicemeshinterface/smi-sdk-go/pkg/apis/specs/v1alpha4"
 	v1alpha20 "github.com/servicemeshinterface/smi-sdk-go/pkg/apis/split/v1alpha2"
 	types "k8s.io/apimachinery/pkg/types"
+	rest "k8s.io/client-go/rest"
 	cache "k8s.io/client-go/tools/cache"
 	v1alpha10 "sigs.k8s.io/mcs-api/pkg/apis/v1alpha1"
 )
@@ -57,6 +58,21 @@ func (m *MockMeshCataloger) AddMeshRootCertificateEventHandler(arg0 cache.Resour
 func (mr *MockMeshCatalogerMockRecorder) AddMeshRootCertificateEventHandler(arg0 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddMeshRootCertificateEventHandler", reflect.TypeOf((*MockMeshCataloger)(nil).AddMeshRootCertificateEventHandler), arg0)
+}
+
+// ConfigFromProxy mocks base method.
+func (m *MockMeshCataloger) ConfigFromProxy(arg0 *models.Proxy, arg1 string, arg2 *rest.Config) (string, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "ConfigFromProxy", arg0, arg1, arg2)
+	ret0, _ := ret[0].(string)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// ConfigFromProxy indicates an expected call of ConfigFromProxy.
+func (mr *MockMeshCatalogerMockRecorder) ConfigFromProxy(arg0, arg1, arg2 interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ConfigFromProxy", reflect.TypeOf((*MockMeshCataloger)(nil).ConfigFromProxy), arg0, arg1, arg2)
 }
 
 // GetEgressClusterConfigs mocks base method.
