+++ pkg/debugger/policy.go
@@ -22,7 +22,7 @@ type policies struct {
 
 func (ds DebugConfig) getOSMConfigHandler() http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		confJSON, err := utils.MeshConfigToJSON(ds.kubeController.GetMeshConfig())
+		confJSON, err := utils.MeshConfigToJSON(ds.computeClient.GetMeshConfig())
 		if err != nil {
 			log.Error().Err(err).Msg("error getting MeshConfig JSON")
 			return
