+++ pkg/debugger/server.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/openservicemesh/osm/pkg/catalog"
 	"github.com/openservicemesh/osm/pkg/certificate"
+	"github.com/openservicemesh/osm/pkg/compute"
 	"github.com/openservicemesh/osm/pkg/envoy/registry"
-	"github.com/openservicemesh/osm/pkg/k8s"
 	"github.com/openservicemesh/osm/pkg/messaging"
 )
 
@@ -42,14 +42,14 @@ func (ds DebugConfig) GetHandlers() map[string]http.Handler {
 // NewDebugConfig returns an implementation of DebugConfig interface.
 func NewDebugConfig(certDebugger *certificate.Manager, xdsDebugger XDSDebugger, meshCatalog catalog.MeshCataloger,
 	proxyRegistry *registry.ProxyRegistry, kubeConfig *rest.Config, kubeClient kubernetes.Interface,
-	kubeController k8s.Controller, msgBroker *messaging.Broker) DebugConfig {
+	computeClient compute.Interface, msgBroker *messaging.Broker) DebugConfig {
 	return DebugConfig{
-		certDebugger:   certDebugger,
-		xdsDebugger:    xdsDebugger,
-		meshCatalog:    meshCatalog,
-		proxyRegistry:  proxyRegistry,
-		kubeClient:     kubeClient,
-		kubeController: kubeController,
+		certDebugger:  certDebugger,
+		xdsDebugger:   xdsDebugger,
+		meshCatalog:   meshCatalog,
+		proxyRegistry: proxyRegistry,
+		kubeClient:    kubeClient,
+		computeClient: computeClient,
 
 		// We need the Kubernetes config to be able to establish port forwarding to the Envoy pod we want to debug.
 		kubeConfig: kubeConfig,
